backend/requirements-dev.txt
# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
Werkzeug==3.0.1

# Database (SQLite for development)
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
# psycopg2-binary==2.9.9  # Uncomment for production/PostgreSQL

# AI/ML Libraries
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
chromadb==0.4.22
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
cryptography==41.0.7

# Payments
stripe==7.10.0

# Utilities
python-multipart==0.0.6
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
pydantic==2.5.3
pydantic-settings==2.1.0

# Email (Optional)
sendgrid==6.11.0

# Monitoring (Optional)
sentry-sdk[flask]==1.39.2

# File Processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2

# Rate Limiting
Flask-Limiter==3.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==22.0.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.7.1
ipython==8.19.0